Introduce write tests and derive func
